refactor: move modules in printer/modules
one level up to printer/
