+++ sketch_map_tool/printer/generate_pdf.py
@@ -16,7 +16,7 @@ from reportlab.pdfgen import canvas
 from svglib.svglib import svg2rlg
 
 from sketch_map_tool.helper_modules.bbox_utils import Bbox
-from sketch_map_tool.printer.modules.paper_formats.paper_formats import (
+from sketch_map_tool.printer.paper_formats.paper_formats import (
     A0,
     A1,
     A2,
