+++ sketch_map_tool/printer/get_map.py
