+++ sketch_map_tool/printer/paper_formats/paper_formats.py
